requirements.txt
# =============================================================================
# requirements.txt (root directory)
# =============================================================================

aws-cdk-lib==2.151.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha==2.151.0a0
boto3>=1.28.0
